♻️ refactor(domain): seperate code
