+++ domain/label_type.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+from domain.validator import NotBlankStr
+
+
+class LabelType(BaseModel):
+    name: NotBlankStr
+
+    class Config:
+        frozen = True
