+++ domain/pos.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+from domain.validator import PositveNumber
+
+
+class Pos(BaseModel):
+    x: PositveNumber
+    y: PositveNumber
+
+    class Config:
+        frozen = True
